edit data type

+++ src/main/java/com/example/screw/vo/EquipmentDataAvgObj.java
@@ -3,21 +3,22 @@ package com.example.screw.vo;
 public class EquipmentDataAvgObj {
 
 	// 資料統計內容 obj: 狀態比例、平均電流、良率
-	private int runRatio;     // 運作的比例
+	private double runRatio;     // 運作的比例
 	
-	private int idleRatio;    // 閒置的比例
+	private double idleRatio;    // 閒置的比例
 	
-	private int errorRaio;    // 錯誤的比例
+	private double errorRaio;    // 錯誤的比例
 	
-	private int currentAvg;   // 一天的平均電流，T 用一天的秒數去計算
+	private double currentAvg;   // 一天的平均電流，T 用一天的秒數去計算
 		
-	private int passRatio;    // 良率
+	private double passRatio;    // 良率
 
 	public EquipmentDataAvgObj() {
 		super();
 	}
 
-	public EquipmentDataAvgObj(int runRatio, int idleRatio, int errorRaio, int currentAvg, int passRatio) {
+	public EquipmentDataAvgObj(double runRatio, double idleRatio, double errorRaio, double currentAvg,
+			double passRatio) {
 		super();
 		this.runRatio = runRatio;
 		this.idleRatio = idleRatio;
@@ -26,43 +27,43 @@ public class EquipmentDataAvgObj {
 		this.passRatio = passRatio;
 	}
 
-	public int getRunRatio() {
+	public double getRunRatio() {
 		return runRatio;
 	}
 
-	public void setRunRatio(int runRatio) {
+	public void setRunRatio(double runRatio) {
 		this.runRatio = runRatio;
 	}
 
-	public int getIdleRatio() {
+	public double getIdleRatio() {
 		return idleRatio;
 	}
 
-	public void setIdleRatio(int idleRatio) {
+	public void setIdleRatio(double idleRatio) {
 		this.idleRatio = idleRatio;
 	}
 
-	public int getErrorRaio() {
+	public double getErrorRaio() {
 		return errorRaio;
 	}
 
-	public void setErrorRaio(int errorRaio) {
+	public void setErrorRaio(double errorRaio) {
 		this.errorRaio = errorRaio;
 	}
 
-	public int getCurrentAvg() {
+	public double getCurrentAvg() {
 		return currentAvg;
 	}
 
-	public void setCurrentAvg(int currentAvg) {
+	public void setCurrentAvg(double currentAvg) {
 		this.currentAvg = currentAvg;
 	}
 
-	public int getPassRatio() {
+	public double getPassRatio() {
 		return passRatio;
 	}
 
-	public void setPassRatio(int passRatio) {
+	public void setPassRatio(double passRatio) {
 		this.passRatio = passRatio;
 	}
