+++ src/main/java/com/example/screw/vo/ProcessObj.java
@@ -4,13 +4,13 @@ public class ProcessObj {   // 單筆製程使用物/排放物的 obj
 	
 	private String name;    // 使用物/排放物的名稱
 	
-	private int amount;     // 使用量/排放量
+	private double amount;     // 使用量/排放量
 
 	public ProcessObj() {
 		super();
 	}
 
-	public ProcessObj(String name, int amount) {
+	public ProcessObj(String name, double amount) {
 		super();
 		this.name = name;
 		this.amount = amount;
@@ -24,11 +24,11 @@ public class ProcessObj {   // 單筆製程使用物/排放物的 obj
 		this.name = name;
 	}
 
-	public int getAmount() {
+	public double getAmount() {
 		return amount;
 	}
 
-	public void setAmount(int amount) {
+	public void setAmount(double amount) {
 		this.amount = amount;
 	}
